+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 ## sys-info-extended
 
+### v0.8.0
+
+- Added `append_env()` and `set_env()` functions for windows. It needs an options argument which has type `EnvOptions` for defining env's name, value and which level of env it is. If you want to set or append an env to System, you have to run the application that your code included as administrator.
+- A bug on `get_language_options()` function debugged, it made public.
+
 ### 0.7.0
 
 - `get_language_options()` function and `LanguageOptions` struct added. It'll take also improvements on next releases.
+- palin, cc and libc versions upgraded.
+
 
 ### 0.6.0